refactor : NameTest와 TurnTest를 InputViewTest로 통합
